se implemento calculo de mascara

+++ proc-de-imagenes/src/main/java/untref/service/ImageConvolutionService.java
@@ -4,5 +4,7 @@ import javafx.scene.image.Image;
 import untref.domain.TemporalColor;
 
 public interface ImageConvolutionService {
-	TemporalColor[][] calculateConvolution(int[][] gausseanLaplacianOperator, Image image, int width, int height);
+	TemporalColor[][] calculateConvolution(double[][] gausseanLaplacianOperator, Image image, int width, int height);
+
+
 }
